Client and Server modules are seperated, ready to build.
