+++ P2P/crypdist-client/src/main/java/Peer.java
@@ -1,6 +1,3 @@
-package client;
-
-
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
