+++ P2P/crypdist-server/src/main/java/Client.java
@@ -1,5 +1,3 @@
-package server;
-
 import java.net.InetAddress;
 
 /**
