+++ P2P/crypdist-server/src/main/java/Main.java
@@ -1,20 +1,17 @@
-package client;
-
-import server.Server;
-
 import java.io.IOException;
 
 /**
  * Created by od on 16.02.2017.
  */
-public class MainClass {
+public class Main {
 
     public static void main (String[] args) {
         try {
-            Thread t = new Client("localhost", 4142, 4544, 4545);
+            Thread t = new Server(4142, 4243);
             t.start();
-        } catch(Exception e) {
+        } catch(IOException e) {
             e.printStackTrace();
         }
     }
+
 }
