+++ P2P/pom.xml
@@ -7,18 +7,11 @@
     <groupId>com.crypdist</groupId>
     <artifactId>com.crypdist.p2p</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
 
+    <packaging>pom</packaging>
 
+    <modules>
+        <module>crypdist-server</module>
+        <module>crypdist-client</module>
+    </modules>
 </project>
